refactor: improve condition parsing logic and enhance test coverage for valid expressions
